Refactor ConnectionConf to allow for database specific options.
ConnectionConf now only has fields that are shared between different databases.
Database specific fields are moved to DbConnectionConf inside the scripting folder.

(cherry picked from commit 5e4661e6595da31ba146c7785eaa911f4a95f547)
