+++ src/scripting/alternator/config.rs
@@ -1,0 +1,5 @@
+use clap::Parser;
+use serde::{Deserialize, Serialize};
+
+#[derive(Parser, Debug, Default, Serialize, Deserialize)]
+pub struct DbConnectionConf {}
