+++ src/scripting/alternator/mod.rs
@@ -1,4 +1,5 @@
 pub mod alternator_error;
+pub mod config;
 pub mod connect;
 pub mod context;
 pub mod functions;
