+++ src/scripting/cql/mod.rs
@@ -1,5 +1,6 @@
 mod bind;
 pub mod cass_error;
+pub mod config;
 pub mod connect;
 pub mod context;
 pub mod cql_types;
